Crate liner

Modules

emacs

vi

Structs

BasicCompleter

Buffer
A buffer for text in the line editor.

Context

Editor
The core line editor. Displays and provides editing for history and the new buffer.

Emacs
Emacs keybindings for Editor. This is the default for Context::read_line().

Event

FilenameCompleter

History
Structure encapsulating command history

Vi
Vi keybindings for Editor.

Enums

Action
A modification performed on a Buffer. These are used for the purpose of undo/redo.

CursorPosition
Represents the position of the cursor relative to words in the buffer.

EventKind

KeyBindings
The key bindings to use.

Traits

Completer

KeyMap

Functions

get_buffer_words
The default for Context.word_divider_fn.

Type Definitions

EventHandler