Enum liner::CursorPosition

pub enum CursorPosition {
    InWord(usize),
    OnWordLeftEdge(usize),
    OnWordRightEdge(usize),
    InSpace(Option<usize>, Option<usize>),
}

Represents the position of the cursor relative to words in the buffer.

Variants
InWord(usize)

The cursor is in the word with the specified index.

OnWordLeftEdge(usize)

The cursor is on the left edge of the word with the specified index. For example: abc |hi, where | is the cursor.

OnWordRightEdge(usize)

The cursor is on the right edge of the word with the specified index. For example: abc| hi, where | is the cursor.

InSpace(Option<usize>, Option<usize>)

The cursor is not in contact with any word. Each Option<usize> specifies the index of the closest word to the left and right, respectively, or None if there is no word on that side.

Methods

impl CursorPosition

pub fn get(cursor: usize, words: &[(usize, usize)]) -> CursorPosition