Struct liner::History

pub struct History {
    pub buffers: VecDeque<Buffer>,
    pub append_duplicate_entries: bool,
    // some fields omitted
}

Structure encapsulating command history

Fields
buffers: VecDeque<Buffer>

Vector of buffers to store history in

append_duplicate_entries: bool
Methods

impl History

pub fn commit_history(&mut self)

It's important to execute this function before exiting your program, as it will ensure that all history data has been written to the disk.

pub fn new() -> History

Create new History structure.

pub fn len(&self) -> usize

Number of items in history.

pub fn push(&mut self, new_item: Buffer) -> Result<()>

Add a command to the history buffer and remove the oldest commands when the max history size has been met. If writing to the disk is enabled, this function will be used for logging history to the designated history file.

pub fn get_newest_match<'a, 'b>(
&'a self,
curr_position: Option<usize>,
new_buff: &'b Buffer
) -> Option<&'a Buffer>

Go through the history and try to find a buffer which starts the same as the new buffer given to this function as argument.

pub fn file_name(&self) -> Option<&str>

Get the history file name.

pub fn set_file_name(&mut self, name: Option<String>)

Set history file name. At the same time enable history.

pub fn set_max_size(&mut self, size: usize)

Set maximal number of buffers stored in memory

pub fn set_max_file_size(&mut self, size: usize)

Set maximal number of entries in history file

pub fn load_history(&mut self) -> Result<()>

Load history from given file name